menyamakan semua bahasa agar konsisten, dan memperbarui tampilan di beberapa fitur

+++ database/migrations/2026_06_22_155303_add_balance_to_users_table.php
@@ -11,11 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (!Schema::hasColumn('users', 'balance')) {
             Schema::table('users', function (Blueprint $table) {
-            $table->decimal('balance', 15, 2)->default(0);
+                $table->decimal('balance', 15, 2)->default(0);
             });
-        });
+        }
     }
 
     /**
@@ -23,10 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'balance')) {
             Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('balance');
+                $table->dropColumn('balance');
             });
-        });
+        }
     }
 };
